WIP: reworking types

+++ posit/fraction.hpp
@@ -27,7 +27,7 @@ public:
 	
 	// selectors
 	bool none() const {	return _Bits.none(); }
-	unsigned int nrBits() const { return _NrOfBits;	}
+	size_t nrBits() const { return _NrOfBits;	}
 	// TODO: this fails when fbits > 64 and we cannot represent the fraction by a 64bit unsigned integer
 	double value() const { return double(_Bits.to_ullong()) / double(uint64_t(1) << fbits);	}
 	quadruple to_quadruple() const { return quadruple(_Bits.to_ullong()) / quadruple(uint64_t(1) << fbits);  }
@@ -204,7 +204,7 @@ private:
 	// maximum size fraction is <nbits - one sign bit - minimum two regime bits>
 	// but we maintain 1 guard bit for rounding decisions
 	std::bitset<fbits> _Bits;
-	unsigned int       _NrOfBits;
+	size_t             _NrOfBits;
 
 	// template parameters need names different from class template parameters (for gcc and clang)
 	// Without the template (i.e. only own operators are friends) we get linker errors
